some warnings solved

+++ caterva/caterva.c
@@ -137,7 +137,7 @@ int caterva_array_append(caterva_context_t *ctx, caterva_array_t *array, void *c
             if (chunksize != array->next_chunksize * array->itemsize) {
                 CATERVA_ERROR(CATERVA_ERR_INVALID_ARGUMENT);
             }
-            CATERVA_ERROR(caterva_blosc_array_append(ctx, array, chunk, chunksize));
+            CATERVA_ERROR(caterva_blosc_array_append(ctx, array, chunk, (int32_t) chunksize));
             break;
         case CATERVA_STORAGE_PLAINBUFFER:
             if (chunksize != array->chunksize * array->itemsize) {
